[DBAL-203] SQLServerPlatform Refactoring
