+++ lib/Doctrine/DBAL/Driver/PDOSqlsrv/Driver.php
@@ -1,7 +1,5 @@
 <?php
 /*
- *  $Id$
- *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
@@ -65,12 +63,12 @@ class Driver implements \Doctrine\DBAL\Driver
 
     public function getDatabasePlatform()
     {
-        return new \Doctrine\DBAL\Platforms\MsSqlPlatform();
+        return new \Doctrine\DBAL\Platforms\SQLServer2005Platform();
     }
 
     public function getSchemaManager(\Doctrine\DBAL\Connection $conn)
     {
-        return new \Doctrine\DBAL\Schema\MsSqlSchemaManager($conn);
+        return new \Doctrine\DBAL\Schema\SQLServerSchemaManager($conn);
     }
 
     public function getName()
