+++ tests/Doctrine/Tests/DBAL/Functional/Schema/SQLServerSchemaManagerTest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Doctrine\Tests\DBAL\Functional\Schema;
+
+use Doctrine\DBAL\Schema;
+
+class SQLServerSchemaManagerTest extends SchemaManagerFunctionalTestCase
+{
+
+}
